Bivariado.SAS
This program obtain the correlations between each variable. The methologies used are Pearson and Spearman tests. 
Once the program gives the results, Pearson test consider -0.6 and 0.6 as the limits for independence betweeen variables while Spearman test consider -0.7 and 0.7 as same limits.
Actually, neither variable have to be correlated. This ensures that all of them are independent.